New app: Yellow Pages
Adds a new app to the phone, Yellow Pages.
You can call a job and it will randomize someone thats on duty and call them.

It's not perfect but it works.

Out of the box it comes with 4 jobs, you can easily add new jobs in yellowpages.js
Not the best design work, cuz i suck at it, so feel free to work on the design and PR it.